+++ dashboard/src/components/settings/signInMethods/WindowsLiveProviderSettings/WindowsLiveProviderSettings.tsx
@@ -13,7 +13,7 @@ import CopyIcon from '@/ui/v2/icons/CopyIcon';
 import Input from '@/ui/v2/Input';
 import InputAdornment from '@/ui/v2/InputAdornment';
 import { copy } from '@/utils/copy';
-import { generateRemoteAppUrl } from '@/utils/helpers';
+import { generateAppServiceUrl } from '@/utils/helpers';
 import { toastStyleProps } from '@/utils/settings/settingsConstants';
 import { FormProvider, useForm } from 'react-hook-form';
 import { toast } from 'react-hot-toast';
@@ -111,9 +111,11 @@ export default function WindowsLiveProviderSettings() {
             fullWidth
             hideEmptyHelperText
             label="Redirect URL"
-            defaultValue={`${generateRemoteAppUrl(
+            defaultValue={`${generateAppServiceUrl(
               currentApplication.subdomain,
-            )}/v1/auth/signin/provider/microsoft/callback`}
+              currentApplication.region.awsName,
+              'auth',
+            )}/v1/signin/provider/windowslive/callback`}
             disabled
             endAdornment={
               <InputAdornment position="end" className="absolute right-2">
@@ -124,9 +126,11 @@ export default function WindowsLiveProviderSettings() {
                   onClick={(e) => {
                     e.stopPropagation();
                     copy(
-                      `${generateRemoteAppUrl(
+                      `${generateAppServiceUrl(
                         currentApplication.subdomain,
-                      )}/v1/auth/signin/provider/microsoft/callback`,
+                        currentApplication.region.awsName,
+                        'auth',
+                      )}/v1/signin/provider/windowslive/callback`,
                       'Redirect URL',
                     );
                   }}