+++ packages/react/src/useSignOut.ts
@@ -1,4 +1,4 @@
-import { signOutPromise } from '@nhost/core'
+import { signOutPromise } from '@nhost/hasura-auth-js'
 import { useSelector } from '@xstate/react'
 
 import { useAuthInterpreter } from './useAuthInterpreter'