+++ examples/codegen-react-apollo/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @nhost-examples/codegen-react-apollo
 
+## 0.1.3
+
+### Patch Changes
+
+- 7f251111: Use `NhostProvider` instead of `NhostReactProvider` and `NhostNextProvider`
+
+  `NhostReactProvider` and `NhostNextProvider` are now deprecated
+
+- Updated dependencies [7f251111]
+  - @nhost/react@0.16.0
+  - @nhost/react-apollo@4.10.0
+
+## 0.1.2
+
+### Patch Changes
+
+- 132a4f4b: chore(deps): synchronize @types/react-dom and @types/react versions
+- Updated dependencies [132a4f4b]
+  - @nhost/react@0.15.2
+  - @nhost/react-apollo@4.9.2
+
 ## 0.1.1
 
 ### Patch Changes