+++ examples/codegen-react-query/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @nhost-examples/codegen-react-query
 
+## 0.1.3
+
+### Patch Changes
+
+- 7f251111: Use `NhostProvider` instead of `NhostReactProvider` and `NhostNextProvider`
+
+  `NhostReactProvider` and `NhostNextProvider` are now deprecated
+
+- Updated dependencies [7f251111]
+  - @nhost/react@0.16.0
+
+## 0.1.2
+
+### Patch Changes
+
+- 132a4f4b: chore(deps): synchronize @types/react-dom and @types/react versions
+- Updated dependencies [132a4f4b]
+  - @nhost/react@0.15.2
+
 ## 0.1.1
 
 ### Patch Changes