+++ examples/nextjs/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @nhost-examples/nextjs
 
+## 0.1.3
+
+### Patch Changes
+
+- 7f251111: Use `NhostProvider` instead of `NhostReactProvider` and `NhostNextProvider`
+
+  `NhostReactProvider` and `NhostNextProvider` are now deprecated
+
+- Updated dependencies [7f251111]
+  - @nhost/nextjs@1.10.0
+  - @nhost/react@0.16.0
+  - @nhost/react-apollo@4.10.0
+
+## 0.1.2
+
+### Patch Changes
+
+- 132a4f4b: chore(deps): synchronize @types/react-dom and @types/react versions
+- Updated dependencies [132a4f4b]
+  - @nhost/react@0.15.2
+  - @nhost/react-apollo@4.9.2
+  - @nhost/nextjs@1.9.3
+
 ## 0.1.1
 
 ### Patch Changes
@@ -9,7 +32,7 @@
 - Updated dependencies [6da44bf8]
   - @nhost/nextjs@1.9.0
   - @nhost/react@0.14.0
-  - @nhost/react-apollo@5.0.0
+  - @nhost/react-apollo@4.10.0
 
 ## 0.0.2