+++ packages/react-apollo/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @nhost/react-apollo
 
+## 4.11.0
+
+### Patch Changes
+
+- Updated dependencies [57db5b83]
+  - @nhost/react@0.17.0
+  - @nhost/apollo@0.5.36
+
+## 4.10.0
+
+### Patch Changes
+
+- Updated dependencies [7f251111]
+  - @nhost/react@0.16.0
+
+## 4.9.2
+
+### Patch Changes
+
+- 132a4f4b: chore(deps): synchronize @types/react-dom and @types/react versions
+- Updated dependencies [132a4f4b]
+  - @nhost/react@0.15.2
+
 ## 4.9.1
 
 ### Patch Changes